+++ Client/src/spadi/controller/ProfitsPercentage.scala
@@ -18,7 +18,8 @@ object ProfitsPercentage
 {
 	// ATTRIBUTES   ----------------------------
 	
-	private val defaultPoints = Vector(Point(0, 23), Point(20, 20), Point(100, 10), Point(200, 5), Point(1000, 1))
+	private val defaultPoints = Vector(Point(0, 23), Point(20, 20), Point(100, 10), Point(200, 5), Point(1000, 2),
+		Point(10000, 1))
 	private val function =
 	{
 		val points =