+++ Client/src/spadi/model/cached/read/PriceKeyMappingFromFieldsFactory.scala
@@ -39,14 +39,30 @@ case class PriceKeyMappingFromFieldsFactory(priceType: PriceType, shopId: Int)
 		case Base => "Perushinta"
 	}
 	
-	override val fieldNames =
+	override def fields =
 	{
-		val base = Vector("Sähkönumero".local -> true, "Nimi".local -> true,
-			"Lisänimi".local -> false, priceFieldName.local -> true, "Yksikkö".local -> true, "Ostomäärä".local -> false)
+		val base = Vector(
+			// Electric id must be 4-11 characters long and must contain at least a single digit
+			InputField.withValidation("Sähkönumero", isRequired = true) { s =>
+				val str = s.getString
+				str.length > 3 && str.length < 12 && str.exists { _.isDigit }
+			},
+			InputField.freeForm("Nimi", isRequired = true),
+			InputField.freeForm("Lisänimi"),
+			// Price must be parseable to a positive double
+			InputField.withValidation(priceFieldName, isRequired = true) { _.double.exists { _ >= 0.0 } },
+			InputField.freeForm("Yksikkö", isRequired = true),
+			// Purchase count must be a positive integer
+			InputField.withValidation("Ostomäärä") { _.int.exists { _ > 0 } }
+		)
 		priceType match
 		{
 			case Net => base
-			case Base => base :+ ("Alennusryhmä".local -> true)
+			// Sale group id must be between 3 and 11 characters
+			case Base => base :+ InputField.withValidation("Alennusryhmä", isRequired = true) { s =>
+				val str = s.getString
+				str.length > 2 && str.length < 12
+			}
 		}
 	}