+++ Client/src/spadi/view/dialog/DataProcessorWindowWithSelections.scala
@@ -3,11 +3,12 @@ package spadi.view.dialog
 import java.nio.file.Path
 
 import spadi.controller.read.DataProcessor
-import spadi.model.cached.read.{KeyMapping, KeyMappingFactory}
+import spadi.model.cached.read.{InputField, KeyMapping, KeyMappingFactory}
 import spadi.model.stored.pricing.Shop
 import spadi.view.component.Fields
 import spadi.view.util.Setup._
 import utopia.flow.datastructure.immutable.Value
+import utopia.flow.datastructure.template.{Model, Property}
 import utopia.flow.generic.ValueConversions._
 import utopia.reflection.component.swing.input.SearchFrom
 
@@ -16,8 +17,10 @@ import utopia.reflection.component.swing.input.SearchFrom
  * @author Mikko Hilpinen
  * @since 7.6.2020, v1.1
  */
-class DataProcessorWindowWithSelections[A, +M <: KeyMapping[A]](path: Path, shop: Shop, mappingFactory: KeyMappingFactory[A, M],
-																headerRow: Vector[String])
+class DataProcessorWindowWithSelections[A, +M <: KeyMapping[A]](path: Path, shop: Shop,
+																mappingFactory: KeyMappingFactory[A, M],
+																headerRow: Vector[String],
+																exampleRows: Vector[Model[Property]])
 										   (makeProcessor: (Path, M) => DataProcessor[A, M])
 	extends DataProcessorWindowLike[A, M, SearchFrom[String, _]](path, shop, mappingFactory)(makeProcessor)
 {
@@ -28,12 +31,13 @@ class DataProcessorWindowWithSelections[A, +M <: KeyMapping[A]](path: Path, shop
 	
 	// IMPLEMENTED  -----------------------------------
 	
-	override protected def keyField(isRequired: Boolean) =
+	override protected def keyField(specification: InputField) =
 	{
 		inputContext.forGrayFields.use { implicit c =>
 			val field = Fields.searchFrom[String](fieldWidth, "Kolumnia nimellä '%s' ei löydetty",
-				if (isRequired) "Hae" else "Hae (vapaaehtoinen)")
-			field.content = headerRow
+				if (specification.isRequired) "Hae" else "Hae (vapaaehtoinen)")
+			// Filters the selectable values based on field validation condition
+			field.content = headerRow.filter { header => specification.allowsValues(exampleRows.map { _(header) }) }
 			field
 		}
 	}